Fix automatic reset into bootloader
The Arduino automatic flashing mechanism triggers a reboot into the
bootloader by setting the UART to 1200 baud and then closing the
connection. The Arduino framework code running on the microcontroller
detects this change, sets a magic value in the last four bytes of SRAM
and resets. The bootloader checks for this magic value and remains in
the bootloader if it's present, otherwise it jumps to the application
code. It's essential for the address and magic value to be in sync
between the application code and the bootloader for this mechanism to
work.

On the SAMD51 Thing Plus, the Arduino framework application code writes
the magic value to 0x2003fffc, but the current bootloader looks for it
at 0x2002fffc. This mismatch causes the automatic bootloader entry to
fail and the bootloader just jumps to application code. Double-tapping
the reset button still works for bootloader entry.

The current bootloader presents a board ID of
SAMD51J19A-SparkFun-Thing-v0 and the 0x2002fffc address corresponds to
4 bytes before the end of 192KB SRAM. The SAMD51 Thing Plus actually
contains a SAMD51J20A with 256KB SRAM. So it appears that the current
bootloader was simply built for the wrong microcontroller part by
mistake.

Unfortuntely it seems like the UF2 board configuration files and
reference to the rest of the bootloader source code has been lost.
Without knowing that exact configuration it's difficult to recreate the
same bootloader binary with the correct microcontroller part. One option
to only fix the automatic flashing mechanism in the current booloader is
to manually patch the DBL_TAP_PTR address. That's what I've done in this
commit.

Following in the footsteps of @shalebridge who started a thread about
this issue [1], I disassembled the bootloader ELF binary with Ghidra,
found three instances of the 0x2002fffc address, and changed them to
the correct 0x2003fffc value. I compared the surrounding code to the
uf2-samdx1 source from roughly the same vintage to verify that I had
found relevant addresses only associated with the automatic update
mechanism (it turns out that all instances of this value are only
associated with the DBL_TAP_PTR mechanism). I also changed the
bootloader version string from "v2.0.0-34-ga8591b7-dirty" to
"v2.0.0-34-ga8591b7-patch" to help with identifying the modified
version.

Next, the modified ELF binary needed to be converted to a raw binary:

  arm-none-eabi-objcopy -O binary bootloader.elf bootloader.bin

That bootloader.bin can be flashed to a board via SWD, but the UF2
bootloader also supports updating itself using a self-update application
that contains an embedded copy of the bootloader. The easiest way to do
that is to just make the same modifications to the existing updater
binary. That is wrapped in a UF2 package, so I used uf2conv.py to
extract the existing updater binary from the UF2 package (this may not
have been necessary, but I didn't investigate the uf2 format):

  python3 uf2conv.py -c -o update-bootloader.bin update-bootloader.uf2

Since update-bootloader.bin contains a copy of the bootloader, I made
all the same modifications to it. It also contained some additional
0x2002fffc address references in the updater code which I updated as
well.

Now there's a catch: the updater binary contains checksums of the
embedded bootloader payload which it uses to verify the bootloader was
flashed correctly. These checksums need to be updated to match the
modified bootloader. During the bootloader build process, gendata.py
converts the bootloader binary to a C byte array and follows that with
another array of checksums over each 1KB of data. These generated arrays
are then normally built into the updater as the payload. I ran
gendata.py twice: once with the original bootloader and again with the
modified bootloader:

  python3 gendata.py 16384 bootloader.bin

I then compared the generated checksums to identify which ones had
changed. I searched for the original checksum values in the updater
binary and replaced them with the new versions.

Now I had a modified updater binary that just needed to be repackaged
in the UF2 format:

  python3 uf2conv.py -c -o update-bootloader.uf2 update-bootloader.bin

To update, double-tap the reset button on the board and copy the
modified update-bootloader.uf2 to the mass-storage device that appears.
It'll reset and the INFO_UF2.TXT file in the bootloader mass-storage
device should display the modified version string. Automatically
entering the bootloader and flashing from the Arduino IDE should now
work correctly as well.

1. https://forum.sparkfun.com/viewtopic.php?f=145&p=228774